requirements.txt
# Backend dependencies
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
openai==1.12.0
python-dotenv==1.0.1
pydantic==2.6.3

# Frontend dependencies
streamlit==1.32.0
requests==2.31.0

# Development and testing
pytest==8.0.2
httpx==0.27.0

# Optional utilities
python-jose==3.3.0  # For JWT handling if needed
passlib==1.7.4      # For password hashing if needed
bcrypt==4.1.2       # For password hashing if needed

# Streamlit extras
streamlit-extras==0.3.6